chore: Fix errors and compiler warnings on verification platforms (#420)
The latest verification run has a few outstanding errors (
https://github.com/apache/arrow-nanoarrow/actions/runs/8584808770/job/23525503382
):

- One file was missing an `#include <array>`
- A Python test error message was hard-coded for little-endian systems

This also fixes a mountain of compiler warnings because of
`-Wmisleading-indentation` on gcc that result from the flatcc generated
header. These warnings made it difficult to find the Python test message
error.

There were also a few other compiler warnings that I noticed when
compiling the tests with Meson. Meson compiles tests with `-Wall` and I
don't think CMake compiles them with anything in particular (we do
compile the C libraries with many warning flags).

+++ src/nanoarrow/schema_test.cc
@@ -1510,7 +1510,8 @@ TEST(MetadataTest, Metadata) {
   std::string simple_metadata = SimpleMetadata();
 
   EXPECT_EQ(ArrowMetadataSizeOf(nullptr), 0);
-  EXPECT_EQ(ArrowMetadataSizeOf(simple_metadata.data()), simple_metadata.size());
+  EXPECT_EQ(ArrowMetadataSizeOf(simple_metadata.data()),
+            static_cast<int64_t>(simple_metadata.size()));
 
   EXPECT_EQ(ArrowMetadataHasKey(simple_metadata.data(), ArrowCharView("key")), 1);
   EXPECT_EQ(ArrowMetadataHasKey(simple_metadata.data(), ArrowCharView("not_a_key")), 0);
